requirements.txt
flask
gunicorn
google-cloud-pubsub
py_translator==1.8.9